% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{percentage_contrib}
\alias{percentage_contrib}
\title{percentage contribution calculation}
\usage{
percentage_contrib(contrib)
}
\arguments{
\item{contrib}{vector of numeric contributions}
}
\description{
A more intuitive way of interpreting contribution, also allows comparison
regardless of resulting total index movement.
}
\details{
See https://github.com/MjStansfi/TPDdecomp for more information
}
