% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPD_decomp.R
\name{TPD_decomp}
\alias{TPD_decomp}
\title{Decomposition of Fixed-Effects index}
\usage{
TPD_decomp(times, logprice, id, weight, custom_time = c(),
  window_length = NULL, verbose = FALSE)
}
\arguments{
\item{times}{vector of the times at which price observations were made. \cr
NOTE: \code{times} must be of class Date or numeric.}

\item{logprice}{vector of log of prices at the given time}

\item{id}{vector of distinct identification number of consumer goods}

\item{weight}{vector of expenditure weights used in the regressions}

\item{custom_time}{either empty (will assume latest periods) or a vector of length
two relating to times compared where [1] is the 'from' time and [2] is the 'to' time
e.g Comparing contribution from "1973-04-01" to "1973-05-01" would be c("1973-04-01", "1973-05-01")}

\item{window_length}{optional. Single number for length of window for the data that
regressions are fit on. Note if window_length is present it assumes custom time is latest two periods
in window.}
}
\value{
A dataframe containing the numeric contribution (\code{contrib}) of every price observation
in the window, the product of which will equal the index movement between the two time periods.

\code{p_contrib} is the standardised percentage contribution of the observation, the sum of which will equal 1.
\code{id_p_contrib} is the standardised percentage contribution of the total id, the sum of which will equal 1.
}
\description{
Take (mostly) the same inputs as the FEWS function and return the decomposition
on the multilateral scale.
}
\examples{
library(TPDdecomp)

load("Turvey.RData")

contributions <- with(
    turvey,
    TPD_decomp(times = month,
               logprice = log(price),
               id = commodity,
               weight = price*quantity,
               custom_time = c("1973-04-30","1973-05-31"),
               window_length = NULL)
 )
#Comparing 1973-05-31 to 1973-04-30

str(contributions)
#Classes ‘data.table’ and 'data.frame':	176 obs. of  7 variables:
#$ times           : Date, format: "1973-12-31" ...
#$ price           : num  1.9 1.57 1.97 2.05 1.36 2.34 2.03 1.08 2.31 2.18 ...
#$ weight          : num  9414 1685 24739 19030 6143 ...
#$ id              : chr  "Apples" "Grapes" "Oranges" "Apples" ...
#$ exp_share       : num  0.263 0.047 0.69 0.415 0.134 ...
#$ contrib         : num  0.999 1 1.001 0.999 1 ...
#$ id_total_contrib: num  1.039 1.017 0.997 1.039 1.017 ...

}
